Create dragon directory in backend. Move dragon.js file over to dragon directory, and rename dragon.js to index.js.
